Refactor into services and a service locator

+++ lib/extensions/barriers_extension.dart
@@ -4,6 +4,7 @@ import '../redux/state/game/barriers.dart';
 
 extension BarriersExtension on Barriers {
   List<Offset> toOffsets() => positions
-      .map<Offset>((position) => Offset(position.x, position.y))
+      .map<Offset>(
+          (position) => Offset(position.x.toDouble(), position.y.toDouble()))
       .toList();
 }
